requirements.txt
fastapi==0.109.1
pydantic
uvicorn[standard]
motor
python-dotenv